% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.medians.R
\name{get.medians}
\alias{get.medians}
\title{Calculate median landmark coordinates based on separate landmark sets}
\usage{
get.medians(AllLMs, outlier.NA = FALSE)
}
\arguments{
\item{AllLMs}{A 4d array that store separate landmark sets with corresponding landmarks and identical samples.}

\item{outlier.NA}{A boolean variable determines whether there is NA (TRUE) or not (FALSE).}
}
\value{
A 3D array that contains median landmark coordinates \cr
}
\description{
This function returns the median landmark estimates based on separate landmark sets.
It is useful to calculate median estimates based on ALPACA estimates of individual templates.
}
\examples{
#Please refer to the help file of 'read.malpaca.estimates' in SlicerMorphR to see how this function calculates median estimates.
}
