% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.markups.json.r
\name{read.markups.json}
\alias{read.markups.json}
\title{Import a JSON file into R}
\usage{
read.markups.json(file = NULL)
}
\arguments{
\item{file}{the file path that leads to a particular JSON file.}
}
\value{
an array that contains the landmark coordinates. The dimensions are specified by dimnames x, y, and z.
}
\description{
This function imports a JSON file into R. It utilizes function "fromJSON" from the package
"jsonlite" to read a JSON file.
}
