% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_lmdists_calc.R
\name{rmse_lmdists_per_specimen}
\alias{rmse_lmdists_per_specimen}
\title{Calculate the root mean square (RMSE) and landmark distances for a specimen (a landmark matrix/array)}
\usage{
rmse_lmdists_per_specimen(Matrix_test, Matrix_ref)
}
\arguments{
\item{Matrix_test}{A matrix or 2d array storing landmark coordinates of a specimen}

\item{Matrix_ref}{A matrix or 2d array storing landmark coordinates of a reference specimen}
}
\value{
A list that contains the RMSE of the specimen as well as pairwise landmark distances \cr
between the specimen and the reference \cr
The returned value contains:
\itemize {
\item $LM_distances = all pairwise landmark distances between a specimen and a reference
\item $RMSE = the RMSE value between a specimen and a reference
}
}
\description{
This function calculates the root mean square (RMSE) and landmark distances for a specimen (a landmark matrix/array)
to a reference specimen (e.g., a mean shape or a manually landmarked specimen)
}
\examples{
results <- rmse_lmdists_per_specimen(specimen1_matrix, mean_matrix)
}
