% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.markups.fcsv.r
\name{write.markups.fcsv}
\alias{write.markups.fcsv}
\title{Export landmark data into FCSV file}
\usage{
write.markups.fcsv(pts = NULL, outfile = NULL)
}
\arguments{
\item{pts}{a matrix that contains landmark coordinates; nrow = number of coordinates; ncol = dimensions}

\item{outfile}{this parameter specifies a file to be created}
}
\value{
a FCSV file of morphometric data
}
\description{
This function export landmark coordinates into an FCSV file
a fcsv file,
}
