% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_parser.R
\name{parser}
\alias{parser}
\title{Parse the GPA log output}
\usage{
parser(file = NULL)
}
\arguments{
\item{file}{an object that points to the location of the analysis.log file, either by hard coding the path or interactively via the function "file.choose()"}
}
\value{
A list that contains every output of the GPA module.
}
\description{
This function parse log output (analysis.log file) from the SlicerMorph's GPA
(General Procrustes analysis) module output.Parser saves everything necessary,
including the output files as well as raw coordinates, in a large R object.
These output files are shown in the analysis.log file.
For details: please see https://github.com/SlicerMorph/Tutorials/blob/main/GPA_3/README.md
}
\examples{
SM.log.file <- file.choose()
SM.log <- parser(SM.log.file)
head(SM.log)
}
