% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.markups.fcsv.r
\name{read.markups.fcsv}
\alias{read.markups.fcsv}
\title{Import a FCSV file into R}
\usage{
read.markups.fcsv(file = NULL)
}
\arguments{
\item{file}{the file path that leads to a particular FCSV file.}
}
\value{
an array that contains the landmark coordinates. The dimensions are specified by dimnames x, y, and z.
}
\description{
This function imports an FCSV file into R. It utilizes the 'read.csv' function to read
a fcsv file,
}
