% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.markups.fcsv.r
\name{read.markups.fcsv}
\alias{read.markups.fcsv}
\title{Import a FCSV file into R}
\usage{
read.markups.fcsv(file = NULL)
}
\arguments{
\item{File}{the file path that leads to a particular FCSV file.}
}
\value{
An array/matrix that contains the landmark coordinates. \cr
Row numbers = number of landmarks;  \cr
Column numbers = dimensions of the landmark coordinates  \cr
Row names are assigned from the labels in the FCSV file  \cr
Column names are assigned as "x", "y", and "z", depending on the dimensionality  \cr
}
\description{
This function imports an FCSV file into R. It utilizes the 'read.csv' function to read
a fcsv file,
}
\examples{
lms = read.markups.fcsv(file = url("https://raw.githubusercontent.com/SlicerMorph/SampleData/master/Gorilla_template_LM1.fcsv"))
#Return a 41x3 matrix; rownames = labels in the FCSV file; colnames = "x", "y", "z"
}
