% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.markups.json.r
\name{read.markups.json}
\alias{read.markups.json}
\title{Import a JSON file into R}
\usage{
read.markups.json(file = NULL)
}
\arguments{
\item{file}{The file path that leads to a particular JSON file.}
}
\value{
An array/matrix that contains the landmark coordinates. \cr
Row numbers = number of landmarks;  \cr
Column numbers = dimensions of the landmark coordinates  \cr
Row names are assigned from the labels in the FCSV file  \cr
Column names are assigned as "x", "y", and "z", depending on the dimension  \cr
}
\description{
This function imports a JSON file into R. It utilizes function "fromJSON" from the package
"jsonlite" to read a JSON file.
}
\examples{
lms = read.markups.json(file = url("https://raw.githubusercontent.com/SlicerMorph/SampleData/master/Gorilla_template_LM1.json"))
#Return a 41x3 matrix; rownames = labels in the FCSV file; colnames = "x", "y", "z"
}
